Abstracted logger out from files now that I know it works.
