+++ lib/passport-azure-ad/bearerstrategy.js
@@ -18,27 +18,12 @@
 
 "use strict";
 
-var bunyan = require('bunyan'),
-BearerStrategy = require('passport-http-bearer').Strategy,
+var BearerStrategy = require('passport-http-bearer').Strategy,
 util = require('util'),
 jwt = require('jsonwebtoken'),
 Metadata = require('./metadata').Metadata,
+Log = require('./logging').getLogger,
 jws = require('jws');
-
-var log = bunyan.createLogger({
-    name: 'Microsoft OpenID Connect: OAuth Bearer Strategy',
-         streams: [
-        {
-            stream: process.stderr,
-            level: "error",
-            name: "error"
-        }, 
-        {
-            stream: process.stdout,
-            level: "warn",
-            name: "console"
-        }, ]
-});
 
 
 var PEMkey = null;
@@ -92,6 +77,8 @@ var PEMkey = null;
 function Strategy(options, verify) {
 /*jshint validthis: true */
 
+    var log = new Log("AzureAD: Bearer Strategy");
+
     // if logging level specified, switch to it.
     if (options.loggingLevel) { log.levels("console", options.loggingLevel); }
 
@@ -115,12 +102,10 @@ if (options.policyName) {
         } 
     }
 
-
-    // if you want to check JWT issuer that is not in the token (not recommended), provide a value here
-
-    if (options.validateIssuer) {
-        log.info('We will validate issuer.');
-    }
+      // warn about validating the issuer
+    if (!options.validateIssuer) {
+      log.warn("We are not validating the issuer. This is fine if you are expecting multiple organizations to connect to your app. Otherwise you should validate the issuer.");
+  }
 
     // if you want to check JWT audience (aud), provide a value here
