+++ lib/passport-azure-ad/metadata.js
@@ -24,24 +24,9 @@ var xml2js = require('xml2js');
 var request = require('request');
 var aadutils = require('./aadutils');
 var async = require('async');
+var Log = require('./logging').getLogger;
 
-// Logging
-
-var bunyan = require('bunyan');
-var log = bunyan.createLogger({
-    name: 'Microsoft OpenID Connect: Passport Strategy: Metadata Parser',
-     streams: [
-        {
-            stream: process.stderr,
-            level: "error",
-            name: "error"
-        }, 
-        {
-            stream: process.stdout,
-            level: "warn",
-            name: "console"
-        }, ]
-});
+var log = new Log("AzureAD: Metadata Parser");
 
 var Metadata = function(url, authtype, options) {
     if (!url) {
@@ -189,7 +174,7 @@ Metadata.prototype.updateOidcMetadata = function(doc, next) {
                     log.info("***********");*/
                     next(null);
                 } catch (e) {
-                    log.warn("No keys found at endpoint!");
+                    log.error("No keys found at endpoint!");
                     next(new Error(e));
                 }
             },
@@ -244,6 +229,7 @@ Metadata.prototype.fetch = function(callback) {
                 if (err) {
                     next(err);
                 } else if (response.statusCode !== 200) {
+                    log.error("Cannot get AAD Federation metadata from endpoint you specified", self.url);
                     next(new Error("Error:" + response.statusCode + " Cannot get AAD Federation metadata from " + self.url));
                 } else {
                     next(null, body);
@@ -268,6 +254,7 @@ Metadata.prototype.fetch = function(callback) {
                 next(null);
 
             } else {
+                log.error("No Authentication type specified to metadata parser. Valid types are saml, wsfed, or odic");
                 next(new Error("No Authentication type specified to metadata parser. Valid types are saml, wsfed, or odic"));
             }
