+++ lib/passport-azure-ad/oidcsetup.js
@@ -18,22 +18,9 @@
  */
 
 var configuration = require('./oidcconfig').configuration,
-bunyan = require('bunyan');
+Log = require('./logging').getLogger;
 
-var log = bunyan.createLogger({
-    name: 'Microsoft OIDC Passport Strategy: Setup',
-         streams: [
-        {
-            stream: process.stderr,
-            level: "error",
-            name: "error"
-        }, 
-        {
-            stream: process.stdout,
-            level: "warn",
-            name: "console"
-        }, ]
-});
+var log = new Log("AzureAD: OIDC Setup");
 
 exports = module.exports = function(identifier, done) {
   log.info('OpenID Discovery...');
