+++ lib/passport-azure-ad/oidcstrategy.js
@@ -28,27 +28,14 @@ var passport = require('passport'),
    OAuth2 = require('oauth').OAuth2,
    setup = require('./oidcsetup'),
    Validator = require('./validator').Validator,
-   bunyan = require('bunyan'),
+   Log = require('./logging').getLogger,
    InternalOAuthError = require('./errors/internaloautherror'),
    Metadata = require('./metadata').Metadata,
    async = require('async');
 
 var cacheManager = require('cache-manager');
 
-var log = bunyan.createLogger({
-    name: 'Microsoft OIDC Passport Strategy',
-     streams: [
-        {
-            stream: process.stderr,
-            level: "error",
-            name: "error"
-        }, 
-        {
-            stream: process.stdout,
-            level: "warn",
-            name: "console"
-        }, ]
-});
+var log = new Log("AzureAD: OIDC Passport Strategy");
 
 var memoryCache = cacheManager.caching({store: 'memory', max: 3600, ttl: 1800/*seconds*/});
 var ttl = 1800; // 30 minutes cache
@@ -85,7 +72,11 @@ function Strategy(options, verify) {
     if (options.loggingLevel) { log.levels("console", options.loggingLevel); }
 
 
+    // warn about validating the issuer
+    if (!options.validateIssuer) {
+      log.warn("We are not validating the issuer. This is fine if you are expecting multiple organizations to connect to your app. Otherwise you should validate the issuer.");
   }
+}
 
 
 /**
@@ -130,7 +121,7 @@ async.waterfall([
     log.info("");
 
     if (!self._options.tenantName) {
-        log.warn("For B2C you must specify a tenant name, none was presented to Strategy as required. (example: tenantName:contoso.onmicrosoft.com");
+        log.error("For B2C you must specify a tenant name, none was presented to Strategy as required. (example: tenantName:contoso.onmicrosoft.com");
         throw new TypeError('OIDCStrategy requires you specify a tenant name to Strategy if using a B2C tenant. (example: tenantName:contoso.onmicrosoft.com');
     }
