+++ lib/passport-azure-ad/tokenValidator.js
@@ -24,26 +24,11 @@
 var jwt = require('jsonwebtoken');
 var jws = require('jws');
 var aadutils = require('./aadutils');
+var Log = require('./logging');
 var PEMkey;
 
 // Logging
-
-var bunyan = require('bunyan');
-var log = bunyan.createLogger({
-    name: 'Microsoft OpenID Connect: Passport Strategy: Token Validator',
-         streams: [
-        {
-            stream: process.stderr,
-            level: "error",
-            name: "error"
-        }, 
-        {
-            stream: process.stdout,
-            level: "warn",
-            name: "console"
-        }, ]
-});
-
+var log = new Log("AzureAD: Token Validation");
 
 var TokenValidator = function(metadata, options) {
     if (!metadata) {
